Cleaned up webcam/camera support.
V4L2 is now the default driver for Linux, and
camiface stays the default for MacOS. 

OpenCV can still be used if V4L2 wrappers are not
installed/found, but it's not as efficient.

Most demos were also cleaned up in terms of 
display: the old qtdisplay() was replaced by
Displayer(), which does not leak memory.